+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.11.1 - 2017-10-11
+
 - [client] Add `EnvHandler::init_with_notify()` to still be notified about
   global events when using `EnvHandler`.
 - [client/server] Externalise state logic to crate `token_store`